requirements.txt
absl-py==2.1.0
gast==0.4.0
grpcio==1.70.0
h5py==3.13.0
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
Markdown==3.7
matplotlib==3.7.1
mock==5.1.0
numpy==1.26.4
opencv-python==4.11.0.86
pillow==10.0.0
protobuf==3.19.6
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
tensorboard==2.10.1
tensorflow==2.10.0
tensorflow-gpu==2.10.0
tensorflow-estimator==2.10.0
termcolor==2.5.0
Werkzeug==3.1.3
seaborn==0.13.2
scikit-learn==1.6.1


flask==2.0.1
flask-cors==3.0.10
pillow==10.0.0